requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.11.0
asyncio==4.0.0
attrs==25.4.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
dotenv==0.9.9
et_xmlfile==2.0.0
fastapi==0.104.1
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
multidict==6.7.0
ngrok==1.5.1
openpyxl==3.1.5
propcache==0.4.0
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.1.1
python-telegram-bot==22.5
redis==6.4.0
requests==2.32.5
sniffio==1.3.1
starlette==0.27.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.37.0
yarl==1.22.0
